dedicated functions for def_* bindings
