+++ rtx_package/src/package/functions.rs
@@ -13,9 +13,14 @@ use rtx_core::common::number::Number;
 use rtx_core::common::store::IntoOption;
 use rtx_core::common::xml::XML_NS;
 use rtx_core::definition::conditional::{Conditional, ConditionalOptions, ConditionalType};
+use rtx_core::definition::constructor::{Constructor, ConstructorOptions};
 use rtx_core::definition::expandable::{Expandable, ExpandableOptions};
+use rtx_core::definition::math_primitive::{MathPrimitive, MathPrimitiveOptions};
+use rtx_core::definition::primitive::{Primitive, PrimitiveOptions};
 use rtx_core::definition::register::{NumericOps, Register, RegisterGetterClosure, RegisterSetterClosure, RegisterType, RegisterValue};
-use rtx_core::definition::{ConditionalClosure, Definition, ExpansionClosure};
+use rtx_core::definition::{
+  BeforeDigestClosure, ConditionalClosure, ConstructionClosure, Definition, DigestionClosure, ExpansionClosure, PrimitiveClosure, ReplacementClosure,
+};
 use rtx_core::document::resource::*;
 use rtx_core::document::tag::{TagOptionName, TagOptions};
 use rtx_core::document::Document;
@@ -25,10 +30,12 @@ use rtx_core::mouth::Mouth;
 use rtx_core::parameter::{Parameter, ParameterExtra, Parameters};
 use rtx_core::state::{Scope, State, Stored};
 use rtx_core::stomach::Stomach;
+use rtx_core::tbox::Tbox;
 use rtx_core::token::{Catcode, Token};
 use rtx_core::tokens::Tokens;
 use rtx_core::util::pathname;
 use rtx_core::util::pathname::PathnameFindOptions;
+use rtx_core::whatsit::Whatsit;
 use rtx_core::BoxOps;
 use rtx_core::{Core, Digested};
 
@@ -805,6 +812,393 @@ pub fn def_register<T: Into<RegisterValue>>(cs: Token, parameters: Option<Parame
     Some(Scope::Global),
   );
   return;
+}
+
+pub fn def_primitive(cs: Token, paramlist: Option<Parameters>, compiled_replacement: PrimitiveClosure, options: PrimitiveOptions, state: &mut State) {
+  let options_locked = options.locked;
+  let scope = options.scope.clone();
+  let mut before_digest_env: Vec<BeforeDigestClosure> = Vec::new();
+  let cs_name = cs.get_cs_name().to_owned();
+
+  if options.require_math {
+    let cs_name_cloned = cs_name.clone();
+    let require_math_closure = beforeproc_single!(stomach, state, { requireMath!(cs_name_cloned, state) });
+    before_digest_env.push(require_math_closure);
+  }
+
+  if options.forbid_math {
+    let cs_name_cloned = cs_name.clone();
+    let forbid_math_closure = beforeproc_single!(stomach, state, { forbidMath!(cs_name_cloned, state) });
+    before_digest_env.push(forbid_math_closure);
+  }
+  if let Some(ref mode) = options.mode {
+    let mode_clone = mode.clone();
+    let begin_mode_closure = beforeproc_single!(stomach, state, {
+      stomach.begin_mode(&mode_clone, state)?;
+    });
+    before_digest_env.push(begin_mode_closure);
+  } else if options.bounded {
+    let bgroup_closure = beforeproc_single!(stomach, state, {
+      stomach.bgroup(state);
+    });
+    before_digest_env.push(bgroup_closure);
+  }
+  if let Some(chosen_font) = options.font {
+    let merge_font_closure = beforeproc_single!(stomach, state, {
+      MergeFont!(&chosen_font, state);
+    });
+    before_digest_env.push(merge_font_closure);
+  }
+  before_digest_env.extend(options.before_digest);
+
+  let mut after_digest_env: Vec<DigestionClosure> = Vec::new();
+  after_digest_env.extend(options.after_digest);
+  if let Some(ref mode) = options.mode {
+    let mode_clone = mode.clone();
+    let end_mode_closure: Vec<DigestionClosure> = afterproc!(stomach, whatsit, state, {
+      stomach.end_mode(&mode_clone, state)?;
+    });
+    after_digest_env.extend(end_mode_closure);
+  } else if options.bounded {
+    let egroup_closure: Vec<DigestionClosure> = afterproc!(stomach, whatsit, state, {
+      stomach.egroup(state)?;
+    });
+    after_digest_env.extend(egroup_closure);
+  }
+
+  state.install_definition(
+    Primitive {
+      cs: cs.clone(),
+      paramlist,
+      replacement: Some(compiled_replacement),
+      options: PrimitiveOptions {
+        before_digest: before_digest_env,
+        after_digest: after_digest_env,
+        ..PrimitiveOptions::default()
+      },
+    },
+    scope,
+  );
+  if options_locked {
+    state.assign_value(&s!("{}:locked", cs_name), true, None);
+  }
+}
+
+pub fn def_math_primitive(cs: Token, paramlist: Option<Parameters>, presentation: String, mut options: MathPrimitiveOptions, state: &mut State) {
+  options.locked = false;
+  options.font = None;
+  let scope = options.scope.clone();
+  let reqfont = options.font.clone().unwrap_or_else(Font::default);
+  state.install_definition(
+    MathPrimitive {
+      cs: cs.clone(),
+      paramlist: None, // never any parameters, this is intentional
+      replacement: Some(Rc::new(move |stomach, args, state| {
+        // let locator    = $stomach->getGullet->getLocator;
+        let mut properties = HashMap::new(); // TODO: sync with perl master here
+        properties.insert(s!("mode"), Stored::String(String::from("math")));
+        // TODO: Improve font precision here, the defaults may not belong in this lookup
+        let font = state
+          .lookup_font()
+          .unwrap_or_else(|| Rc::new(Font::default()))
+          .merge(&reqfont)
+          .specialize(&presentation);
+        let font = Rc::new(font);
+        // foreach my $key (keys %properties) {
+        //   my $value = $properties{$key};
+        //   if (ref $value eq 'CODE') {
+        //     $properties{$key} = &$value(); } }
+        // info!("defmath_prim: {}, tokens: {:?}", &$presentation, $cs);
+        Ok(vec![Digested::TBox(Rc::new(
+          // TODO: Can we reduce boilerplate?
+          Tbox {
+            text: presentation.clone(),
+            tokens: Tokens!(cs.clone()),
+            font,
+            properties,
+            ..Tbox::default()
+          },
+        ))])
+      })),
+      options,
+      ..MathPrimitive::default()
+    },
+    scope,
+  );
+}
+
+pub fn def_constructor(
+  cs: Token,
+  paramlist: Option<Parameters>,
+  compiled_replacement: Option<ReplacementClosure>,
+  options: ConstructorOptions,
+  state: &mut State,
+)
+{
+  // TODO: This won't work, as we can only invoke method calls on paramlist in runtime
+  //*rtx_codegen::constructable::NARGS = $paramlist.get_num_args();
+  let scope = options.scope.clone();
+  let is_locked = options.locked.clone();
+  let cs_name = cs.get_cs_name().to_owned();
+  let locked_key = if is_locked { s!("{}:locked", cs_name) } else { String::new() };
+
+  let mut before_digest_closures: Vec<BeforeDigestClosure> = Vec::new();
+
+  if options.require_math {
+    let cs_name_cloned = cs_name.clone();
+    let require_math_closure = beforeproc_single!(stomach, state, { requireMath!(cs_name_cloned, state) });
+    before_digest_closures.push(require_math_closure);
+  }
+  if options.forbid_math {
+    let cs_name_cloned = cs_name.clone();
+    let forbid_math_closure = beforeproc_single!(stomach, state, { forbidMath!(cs_name_cloned, state) });
+    before_digest_closures.push(forbid_math_closure);
+  }
+  if let Some(ref mode) = options.mode {
+    let mode_clone = mode.clone();
+    let begin_mode_closure = beforeproc_single!(stomach, state, {
+      stomach.begin_mode(&mode_clone, state)?;
+    });
+    before_digest_closures.push(begin_mode_closure);
+  } else if options.bounded {
+    let bgroup_closure = beforeproc_single!(stomach, state, {
+      stomach.bgroup(state);
+    });
+    before_digest_closures.push(bgroup_closure);
+  }
+  if let Some(chosen_font) = options.font {
+    let merge_font_closure = beforeproc_single!(stomach, state, {
+      MergeFont!(&chosen_font, state);
+    });
+    before_digest_closures.push(merge_font_closure);
+  }
+  before_digest_closures.extend(options.before_digest);
+
+  let mut after_digest_closures: Vec<DigestionClosure> = Vec::new();
+  after_digest_closures.extend(options.after_digest);
+  if let Some(ref mode) = options.mode {
+    let mode_clone = mode.clone();
+    let end_mode_closure: Vec<DigestionClosure> = afterproc!(stomach, whatsit, state, {
+      stomach.end_mode(&mode_clone, state)?;
+    });
+    after_digest_closures.extend(end_mode_closure);
+  } else if options.bounded {
+    let egroup_closure: Vec<DigestionClosure> = afterproc!(stomach, whatsit, state, {
+      stomach.egroup(state)?;
+    });
+    after_digest_closures.extend(egroup_closure);
+  }
+
+  let constructor = Constructor {
+    cs,
+    paramlist,
+    replacement: compiled_replacement,
+    before_digest: before_digest_closures,
+    after_digest: after_digest_closures,
+    before_construct: options.before_construct,
+    after_construct: options.after_construct,
+    nargs: options.nargs,
+    alias: options.alias,
+    reversion: options.reversion,
+    // sizer
+    capture_body: options.capture_body,
+    properties: options.properties,
+    // outer
+    // long
+    ..Constructor::default()
+  };
+  state.install_definition(constructor, scope);
+
+  if is_locked {
+    state.assign_value(&locked_key, true, None);
+  }
+}
+
+pub fn def_environment(
+  name: String,
+  paramlist: Option<Parameters>,
+  compiled_replacement: Option<ReplacementClosure>,
+  options: ConstructorOptions,
+  state: &mut State,
+)
+{
+  let begin_name = s!("\\begin{{{}}}", &name);
+  let end_name = s!("\\end{{{}}}", &name);
+  // This is for the common case where the environment is opened by \begin{env}
+  // let sizer = inferSizer($options.sizer, $options.reversion);
+  let mut before_digest_env: Vec<BeforeDigestClosure> = Vec::new();
+  match &options.mode {
+    Some(ref mode) => {
+      let bmode = mode.clone();
+      let mode_closure = Rc::new(move |stomach: &mut Stomach, state: &mut State| {
+        stomach.begin_mode(&bmode, state)?;
+        Ok(Vec::new())
+      });
+      before_digest_env.push(mode_closure);
+    },
+    None => {
+      let bgroup_closure = beforeproc_single!(stomach, state, {
+        stomach.bgroup(state);
+      });
+      before_digest_env.push(bgroup_closure);
+    },
+  };
+  if options.require_math {
+    let require_name = begin_name.clone();
+    let require_math_closure = beforeproc_single!(stomach, state, { requireMath!(require_name, state) });
+    before_digest_env.push(require_math_closure);
+  }
+  if options.forbid_math {
+    let forbid_name = begin_name.clone();
+    let forbid_math_closure = beforeproc_single!(stomach, state, { forbidMath!(forbid_name, state) });
+    before_digest_env.push(forbid_math_closure);
+  }
+
+  let env_name = name.clone();
+  let current_environment_closure = beforeproc_single!(stomach, state, {
+    AssignValue!("current_environment", env_name.clone(), None, state);
+    let body = T_LETTER!(env_name.clone());
+    DefMacroI!(T_CS!("\\@currenvir"), None, body.clone(), state);
+  });
+  before_digest_env.push(current_environment_closure);
+
+  if let Some(chosen_font) = options.font {
+    let merge_font_closure = beforeproc_single!(stomach, state, {
+      MergeFont!(&chosen_font.clone(), state);
+    });
+    before_digest_env.push(merge_font_closure);
+  }
+  before_digest_env.extend(options.before_digest);
+
+  let push_frame_closure = Rc::new(|_document: &mut Document, _whatsit: &Whatsit, state: &mut State| {
+    state.push_frame();
+    Ok(())
+  });
+  let mut before_construct_with_frame: Vec<ConstructionClosure> = vec![push_frame_closure];
+  before_construct_with_frame.extend(options.before_construct);
+
+  let mut after_construct_with_frame: Vec<ConstructionClosure> = options.after_construct;
+
+  let pop_frame_closure = Rc::new(|_document: &mut Document, _whatsit: &Whatsit, state: &mut State| {
+    state.pop_frame()?;
+    Ok(())
+  });
+  after_construct_with_frame.push(pop_frame_closure);
+
+  let begin_name_constructor = Rc::new(Constructor {
+    cs: T_CS!(begin_name),
+    paramlist: paramlist.clone(),
+    replacement: compiled_replacement.clone(),
+    nargs: options.nargs,
+    before_digest: before_digest_env,
+    after_digest: options.after_digest_begin,
+    after_digest_body: options.after_digest_body,
+    before_construct: before_construct_with_frame,
+    // Curiously, it's the \begin whose afterConstruct gets called.
+    after_construct: after_construct_with_frame,
+    capture_body: true,
+    properties: options.properties.clone(),
+    // (defined $options{reversion} ? (reversion => $options{reversion}) : ()),
+    // (defined $sizer ? (sizer => $sizer) : ()),
+    // ), $options{scope});
+    reversion: options.reversion,
+    alias: options.alias,
+    ..Constructor::default()
+  });
+  state.install_definition(begin_name_constructor, options.scope.clone());
+
+  let mut after_digest_env = options.after_digest;
+  let unexpected_end_closure = Rc::new(|_stomach: &mut Stomach, _whatsit: &mut Whatsit, state: &mut State| {
+    // let env = LookupValue!("current_environment", $state_arg);
+    //     Error('unexpected', "\\end{$name}", $_[0],
+    //       "Can't close environment $name",
+    //       "Current are "
+    //         . join(', ', state->lookupStackedValues('current_environment')))
+    //       unless $env && $name eq $env;
+    //     return; },
+    Ok(Vec::new())
+  });
+  after_digest_env.push(unexpected_end_closure);
+
+  match options.mode {
+    Some(mode) => {
+      let emode = mode.clone();
+      let emode_closure = Rc::new(move |stomach: &mut Stomach, _whatsit: &mut Whatsit, state: &mut State| {
+        stomach.end_mode(&emode, state)?;
+        Ok(Vec::new())
+      });
+      after_digest_env.push(emode_closure);
+    },
+    None => {
+      let egroup_closure = Rc::new(|stomach: &mut Stomach, _whatsit: &mut Whatsit, state: &mut State| {
+        stomach.egroup(state)?;
+        Ok(Vec::new())
+      });
+      after_digest_env.push(egroup_closure);
+    },
+  };
+
+  let end_envname_constructor = Rc::new(Constructor {
+    cs: T_CS!(end_name),
+    replacement: None,
+    paramlist: None,
+    before_digest: options.before_digest_end,
+    after_digest: after_digest_env,
+    ..Constructor::default() // TODO ? fill in missing ones
+  });
+  state.install_definition(end_envname_constructor, options.scope.clone());
+
+  // For the uncommon case opened by \csname env\endcsname
+  let name_constructor = Rc::new(Constructor {
+    cs: T_CS!(s!("\\{}", &name)),
+    paramlist,
+    replacement: compiled_replacement,
+    // beforeDigest => flatten(($options{requireMath} ? (sub { requireMath($name); }) : ()),
+    //   ($options{forbidMath} ? (sub { forbidMath($name); })              : ()),
+    //   ($mode                ? (sub { $_[0]->beginMode($mode); })        : ()),
+    //   ($options{font}       ? (sub { MergeFont(%{ $options{font} }); }) : ()),
+    //   $options{beforeDigest}),
+    // afterDigest     => flatten($options{afterDigestBegin}),
+    // afterDigestBody => flatten($options{afterDigestBody}),
+    // beforeConstruct => flatten(sub { state->pushFrame; }, $options{beforeConstruct}),
+    // Curiously, it's the \begin whose afterConstruct gets called.
+    // afterConstruct => flatten($options{afterConstruct}, sub { state->popFrame; }),
+    nargs: options.nargs,
+    capture_body: true,
+    properties: options.properties.clone(),
+    // (defined $options{reversion} ? (reversion => $options{reversion}) : ()),
+    // (defined $sizer ? (sizer => $sizer) : ()),
+    // ), $options{scope});
+    ..Constructor::default()
+  });
+  state.install_definition(name_constructor, options.scope.clone());
+
+  let end_name_constructor = Rc::new(Constructor {
+    cs: T_CS!(s!("\\end{}", &name)),
+    paramlist: None,
+    replacement: Some(Rc::new(|document, whatsit, properties, state| {
+      let env = state.lookup_value("current_environment");
+      // Error('unexpected', "\\end{$name}", $_[0],
+      //   "Can't close environment $name",
+      //   "Current are "
+      //     . join(', ', state->lookupStackedValues('current_environment')))
+      //   unless $env && $name eq $env;
+      Ok(())
+    })),
+    // beforeDigest => flatten($options{beforeDigestEnd}),
+    // afterDigest  => flatten($options{afterDigest},
+    //   ($mode ? (sub { $_[0]->endMode($mode); }) : ())),
+    // ), $options{scope});
+    ..Constructor::default()
+  });
+  state.install_definition(end_name_constructor, options.scope);
+
+  if options.locked {
+    state.assign_value(&s!("\\begin{{{}}}:locked", &name), true, None);
+    state.assign_value(&s!("\\end{{{}}}:locked", &name), true, None);
+    state.assign_value(&s!("\\{}:locked", &name), true, None);
+    state.assign_value(&s!("\\end{}:locked", &name), true, None);
+  }
 }
 
 //**********************************************************************
