+++ rtx_package/src/package/pool/latex_math_mode_environments.rs
@@ -1,4 +1,3 @@
-use std::borrow::Cow;
 use std::rc::Rc;
 
 use crate::package::*;
